support defRaft

+++ alley_oop/interpol/gp_warpfield.py
@@ -68,7 +68,7 @@ class GP_WarpFieldEstimator(torch.nn.Module):
         def_field_se3 = target_opts - ref_opts
 
         # map the depth confidence to a noise prior
-        noise = torch.ones_like(def_field_se3) if noise is None else noise
+        noise = torch.ones_like(def_field_se3, dtype=torch.float16) if noise is None else noise
         self.gp_regressor.fit(ref_opts.T, def_field_se3, noise_level=noise*self.noise_level)
 
     def predict(self, ref_opts: torch.tensor):
